Download File dan Agenda

+++ src/App/Agenda/Controller/AgendaController.php
@@ -30,34 +30,39 @@ class AgendaController extends GlobalFunc
 
         return $this->render_template('agenda/edit', ['idAgenda' => $datas['idAgenda'], 'namaAgenda'=>$datas['namaAgenda'], 'deskripsiAgenda'=>$datas['deskripsiAgenda'], 'datestartAgenda'=>$datas['datestartAgenda'], 'dateendAgenda'=>$datas['dateendAgenda']]);
     }
-    public function update(Request $request)
-    {
-        $id = $request->request->get('id');
-        $namaAgenda = $request->request->get('namaAgenda');
-        $deskripsiAgenda = $request->request->get('deskripsiAgenda');
-        $datestartAgenda = $request->request->get('datestartAgenda');
-        $dateendAgenda = $request->request->get('dateendAgenda');
-
-        $id = $request->attributes->get('id');
-        $data_test = array(
-            'namaAgenda' => $namaAgenda,
-            'deskripsiAgenda' => $deskripsiAgenda,
-            'datestartAgenda' => $datestartAgenda,
-            'dateendAgenda' => $dateendAgenda
-        );
-        
-       $this->model->update($id, $data_test);
-
-       return header("location:http://kesbangpol.com/agenda");
-    }
-    public function delete(Request $request)
+    public function updateordelete(Request $request)
     {
         $id = $request->attributes->get('id');
-        $this->model->delete($id);
+        $delete = $request->request->get('delete');
 
+        if ($delete == null){
+            $namaAgenda = $request->request->get('namaAgenda');
+            $deskripsiAgenda = $request->request->get('deskripsiAgenda');
+            // $datestartAgenda = $request->request->get('datestartAgenda');
+            // $dateendAgenda = $request->request->get('dateendAgenda');
+
+            $data_test = array(
+                'namaAgenda' => $namaAgenda,
+                'deskripsiAgenda' => $deskripsiAgenda,
+                // 'datestartAgenda' => $datestartAgenda,
+                // 'dateendAgenda' => $dateendAgenda
+            );
+            
+            $this->model->update($id, $data_test);
+        } else {
+            $this->model->delete($id);
+        }
 
         return header("location:http://kesbangpol.com/agenda");
     }
+    // public function delete(Request $request)
+    // {
+    //     $id = $request->attributes->get('id');
+    //     $this->model->delete($id);
+
+
+    //     return header("location:http://kesbangpol.com/agenda");
+    // }
     public function create(Request $request)
     {
         return $this->render_template('agenda/create');
